Remove deprecated query documents code
