+++ src/QueryDocuments/QueryBuilderProperty.spec.ts
@@ -1,28 +1,21 @@
-import { OptionalToken, VariableToken } from "sparqler/tokens";
+import { VariableToken } from "sparqler/tokens";
 
-import {
-	createMockContext,
-	createMockDigestedSchema,
-	createMockDigestedSchemaProperty,
-	MockQueryContainer
-} from "../../test/helpers/mocks";
+import { createMockContext, MockQueryContainer } from "../../test/helpers/mocks";
 
 import { AbstractContext } from "../Context/AbstractContext";
 
-import { DigestedObjectSchema } from "../ObjectSchema/DigestedObjectSchema";
 import { DigestedObjectSchemaProperty } from "../ObjectSchema/DigestedObjectSchemaProperty";
 
-import { clazz, constructor, hasProperty, hasSignature, INSTANCE, method, module, } from "../test/JasmineExtender";
+import { clazz, constructor, hasProperty, hasSignature, INSTANCE, module, } from "../test/JasmineExtender";
 
 import { QueryBuilderProperty } from "./QueryBuilderProperty";
 import { QueryContainer } from "./QueryContainer";
-import { QueryPropertyType } from "./QueryPropertyType";
 
 
-describe( module( "carbonldp/QueryDocuments/QueryProperty2" ), ():void => {
+describe( module( "carbonldp/QueryDocuments/QueryBuilderProperty" ), ():void => {
 
 	describe( clazz(
-		"CarbonLDP.QueryDocuments.QueryProperty2",
+		"CarbonLDP.QueryDocuments.QueryBuilderProperty",
 		"Class that represents a property in the query"
 	), ():void => {
 
@@ -150,130 +143,19 @@ describe( module( "carbonldp/QueryDocuments/QueryProperty2" ), ():void => {
 		} );
 
 
-		// TODO: Test .setType
-
-		// TODO: Test .isEmpty
-		// TODO: Test .isPartial
-		// TODO: Test .isComplete
-
-
-		// TODO: Test .getTriple
-
-		describe( method( INSTANCE, "getSearchPatterns" ), ():void => {
-
-			it( hasSignature(
-				"Returns the pattern to be used in the query that specifies the property and its elements\n" +
-				"If the property is optional the patterns will be wrapped in an optional SPARQL token.",
-				{ type: "sparqler/tokens/PatternToken[]" }
-			), ():void => {} );
-
-			it( "should exists", ():void => {
-				expect( QueryBuilderProperty.prototype.getSearchPatterns ).toBeDefined();
-				expect( QueryBuilderProperty.prototype.getSearchPatterns ).toEqual( jasmine.any( Function ) );
-			} );
-
-			it( "should return the patterns as optional", ():void => {
-				const queryProperty:QueryBuilderProperty = new QueryBuilderProperty( {
-					queryContainer: queryContainer,
-					name: "name",
-					parent: new QueryBuilderProperty( {
-						queryContainer: queryContainer,
-						name: "parent",
-						definition: new DigestedObjectSchemaProperty(),
-						optional: true,
-					} ),
-					definition: new DigestedObjectSchemaProperty(),
-					optional: true,
-				} );
-
-				expect( queryProperty.getSearchPatterns() ).toEqual( [
-					jasmine.any( OptionalToken ) as any as OptionalToken,
-				] );
-			} );
-
-			it( "should return the patterns when obligatory", ():void => {
-				const queryProperty:QueryBuilderProperty = new QueryBuilderProperty( {
-						queryContainer: queryContainer,
-						name: "name",
-						parent: new QueryBuilderProperty( {
-							queryContainer: queryContainer,
-							name: "parent",
-							definition: new DigestedObjectSchemaProperty(),
-							optional: true,
-						} ),
-						definition: new DigestedObjectSchemaProperty(),
-						optional: true,
-					} )
-				;
-
-				queryProperty.setObligatory();
-
-				expect( queryProperty.getSearchPatterns() ).not.toEqual( [
-					jasmine.any( OptionalToken ) as any as OptionalToken,
-				] );
-			} );
-
-			// TODO: Test more cases
-
-		} );
-
+		// TODO: Test .getSelfPattern
+		// TODO: Test .getSearchPatterns
 		// TODO: Test .getConstructPatterns
 
 
 		// TODO: Test .hasProperties
-		// TODO: Test .getRootProperty
 		// TODO: Test .addProperty
+
 
 		// TODO: Test .addValues
 		// TODO: Test .addType
+		// TODO: Test .addFilter
 		// TODO: Test .setObligatory
-
-		describe( method( INSTANCE, "getSchema" ), ():void => {
-
-			it( hasSignature(
-				"Returns the specific schema for the property objects that was created query definition.",
-				{ type: "CarbonLDP.DigestedObjectSchema" }
-			), ():void => {} );
-
-			it( "should exists", ():void => {
-				expect( QueryBuilderProperty.prototype.getSchemaFor ).toBeDefined();
-				expect( QueryBuilderProperty.prototype.getSchemaFor ).toEqual( jasmine.any( Function ) );
-			} );
-
-			it( "should return empty schema when no properties", ():void => {
-				const queryProperty:QueryBuilderProperty = new QueryBuilderProperty( {
-					queryContainer: queryContainer,
-					name: "name",
-					definition: new DigestedObjectSchemaProperty(),
-					optional: true,
-				} );
-
-				const propertySchema:DigestedObjectSchema = queryProperty.getSchemaFor( {} );
-				expect( propertySchema ).toEqual( new DigestedObjectSchema() );
-			} );
-
-			it( "should return schema from sub-properties", ():void => {
-				const queryProperty:QueryBuilderProperty = new QueryBuilderProperty( {
-					queryContainer: queryContainer,
-					name: "name",
-					definition: new DigestedObjectSchemaProperty(),
-					optional: true,
-					propertyType: QueryPropertyType.PARTIAL,
-				} );
-
-				queryProperty.addProperty( "property1", {} );
-				queryProperty.addProperty( "property2", { "@id": "https://example.com/property2" } );
-
-				const propertySchema:DigestedObjectSchema = queryProperty.getSchemaFor( {} );
-				expect( propertySchema ).toEqual( createMockDigestedSchema( {
-					properties: new Map<string, DigestedObjectSchemaProperty>( [
-						[ "property1", createMockDigestedSchemaProperty( { uri: "https://example.com/vocab#property1" } ) ],
-						[ "property2", createMockDigestedSchemaProperty( { uri: "https://example.com/property2" } ) ],
-					] ),
-				} ) );
-			} );
-
-		} );
 
 	} );
