+++ src/QueryDocuments/QueryDocumentsBuilder.ts
@@ -1,100 +1,41 @@
-import { LimitToken, OffsetToken, OptionalToken, OrderToken, SubjectToken, SubSelectToken } from "sparqler/tokens";
-
 import { IllegalArgumentError } from "../Errors/IllegalArgumentError";
-import { IllegalStateError } from "../Errors/IllegalStateError";
 
-import { QueryDocumentBuilder } from "./QueryDocumentBuilder";
-import { QueryProperty } from "./QueryProperty";
-import { _getParentPath } from "./Utils";
+import { SubQueryDocumentsBuilder } from "./QueryDocumentBuilder";
+import { QueryDocumentContainer } from "./QueryDocumentContainer";
+import { QueryRootProperty } from "./QueryRootProperty";
 
 
-export interface QueryDocumentsBuilderOrderData {
-	path:string;
-	flow?:"ASC" | "DESC";
-}
+export class QueryDocumentsBuilder extends SubQueryDocumentsBuilder {
+	readonly _queryProperty:QueryRootProperty;
 
-export class QueryDocumentsBuilder extends QueryDocumentBuilder {
+	constructor( queryContainer:QueryDocumentContainer, queryProperty:QueryRootProperty ) {
+		super( queryContainer, queryProperty );
+	}
 
-	_orderData?:QueryDocumentsBuilderOrderData;
 
 	orderBy( property:string, flow?:"ASC" | "DESC" | "ascending" | "descending" ):this {
-		let propertyObj:QueryProperty = this.property( property );
-
-		const select:SubSelectToken = this.__getSubSelect();
-
-		this._orderData = void 0;
-		const orderIndex:number = select.modifiers.findIndex( pattern => pattern.token === "order" );
-
-		if( orderIndex !== - 1 ) {
-			select.modifiers.splice( orderIndex, 1 );
-
-			const optionalIndex:number = select.where.groupPattern.patterns.findIndex( pattern => pattern.token === "optional" );
-			select.where.groupPattern.patterns.splice( optionalIndex, 1 );
-		}
-
-		const validatedFlow:"ASC" | "DESC" = parseFlowString( flow );
-		select.modifiers.unshift( new OrderToken( propertyObj.variable, validatedFlow ) );
-
-		const orderData:QueryDocumentsBuilderOrderData = {
-			path: propertyObj.name
-				.split( "." )
-				.slice( 1 )
-				.join( "." ),
-			flow: validatedFlow,
-		};
-
-		let propertyPatternsPath:OptionalToken;
-		while( propertyObj !== this._document ) {
-			const propertyTriple:SubjectToken = propertyObj && propertyObj.getTriple();
-			if( ! propertyTriple ) throw new IllegalArgumentError( `The property "${ propertyObj.name }" is not a valid property defined by the builder.` );
-
-			const propertyPattern:OptionalToken = new OptionalToken()
-				.addPattern( propertyTriple );
-
-			if( propertyPatternsPath ) propertyPattern.addPattern( propertyPatternsPath );
-			propertyPatternsPath = propertyPattern;
-
-			propertyObj = this._context.getProperty( _getParentPath( propertyObj.name ) );
-		}
-
-		this._orderData = orderData;
-		select.addPattern( propertyPatternsPath );
+		this._queryProperty.setOrder( {
+			path: property,
+			flow: parseFlowString( flow ),
+		} );
 
 		return this;
 	}
 
 	limit( limit:number ):this {
-		const select:SubSelectToken = this.__getSubSelect();
-
-		const limitIndex:number = select.modifiers.findIndex( pattern => pattern.token === "limit" );
-		if( limitIndex !== - 1 ) select.modifiers.splice( limitIndex, 1 );
-
-		select.modifiers.push( new LimitToken( limit ) );
+		this._queryProperty.setLimit( limit );
 
 		return this;
 	}
 
 	offset( offset:number ):this {
-		const select:SubSelectToken = this.__getSubSelect();
+		this._queryProperty.setOffset( offset );
 
-		const offsetIndex:number = select.modifiers.findIndex( pattern => pattern.token === "offset" );
-		if( offsetIndex !== - 1 ) select.modifiers.splice( offsetIndex, 1 );
-
-		select.modifiers.push( new OffsetToken( offset ) );
 		return this;
 	}
 
-
-	private __getSubSelect():SubSelectToken {
-		const select:SubSelectToken = this._document
-			.getPatterns()
-			.find( ( pattern ):pattern is SubSelectToken => pattern.token === "subSelect" );
-
-		if( ! select ) throw new IllegalStateError( `A sub-select token has not been defined.` );
-		return select;
-	}
-
 }
+
 
 function parseFlowString( flow?:"ASC" | "DESC" | "ascending" | "descending" ):"ASC" | "DESC" {
 	if( flow === void 0 ) return void 0;
@@ -116,4 +57,3 @@ function parseFlowString( flow?:"ASC" | "DESC" | "ascending" | "descending" ):"A
 			throw new IllegalArgumentError( "Invalid flow order." );
 	}
 }
-
