+++ src/QueryDocuments/QueryObject.ts
@@ -1,20 +1,21 @@
 import { isBNodeLabel } from "sparqler/iri";
 import { BlankNodeToken, IRIToken, PrefixedNameToken } from "sparqler/tokens";
 
-import { Pointer } from "../Pointer/Pointer";
+import { QueryDocumentContainer } from "./QueryDocumentContainer";
 
-import { isString } from "../Utils";
-import { QueryContext } from "./QueryContext";
 
 export class QueryObject {
-	private _context:QueryContext;
-	private _resource:IRIToken | BlankNodeToken | PrefixedNameToken;
+	private readonly _queryContainer:QueryDocumentContainer;
+	private readonly _resource:IRIToken | BlankNodeToken | PrefixedNameToken;
 
-	constructor( context:QueryContext, object:Pointer | string ) {
-		this._context = context;
-		const id:string = isString( object ) ? object : object.$id;
-		this._resource = isBNodeLabel( id ) ? new BlankNodeToken( id ) : this._context.compactIRI( id );
+	constructor( queryContainer:QueryDocumentContainer, id:string ) {
+		this._queryContainer = queryContainer;
+
+		this._resource = isBNodeLabel( id )
+			? new BlankNodeToken( id )
+			: this._queryContainer.compactIRI( id );
 	}
+
 
 	getToken():IRIToken | BlankNodeToken | PrefixedNameToken {
 		return this._resource;
