--- src/QueryDocuments/QuerySchema2.ts
@@ -1,6 +1,0 @@
-import { QuerySchemaProperty2 } from "./QuerySchemaProperty2";
-
-
-export interface QuerySchema2 {
-	[ propertyName:string ]:QuerySchemaProperty2 | string;
-}
