+++ src/QueryDocuments/QuerySchemaProperty.ts
@@ -1,9 +1,13 @@
-import { QueryDocumentBuilder } from "./QueryDocumentBuilder";
+import { Path, PathBuilder } from "sparqler/patterns";
+
+import { SubQueryDocumentsBuilder } from "./QueryDocumentBuilder";
+
 
 export interface QuerySchemaProperty {
 	"@id"?:string;
 	"@type"?:"@id" | string;
 	"@language"?:string;
 	"@container"?:"@set" | "@list" | "@language";
-	query?:( queryBuilder:QueryDocumentBuilder ) => QueryDocumentBuilder;
+	query?:( queryBuilder:SubQueryDocumentsBuilder ) => SubQueryDocumentsBuilder;
+	path?:( pathBuilder:PathBuilder ) => Path;
 }
