+++ src/QueryDocuments/SubQueryPropertyDefinition.ts
@@ -1,5 +1,6 @@
-import { QuerySchemaProperty2 } from "./QuerySchemaProperty2";
+import { QuerySchemaProperty } from "./QuerySchemaProperty";
 
-export interface SubQueryPropertyDefinition extends QuerySchemaProperty2 {
+
+export interface SubQueryPropertyDefinition extends QuerySchemaProperty {
 	inherit?:false;
 }
