+++ src/QueryDocuments/Utils.spec.ts
@@ -1,22 +1,4 @@
-import {
-	GraphToken,
-	IRIRefToken,
-	OptionalToken,
-	PatternToken,
-	PropertyToken,
-	SubjectToken,
-	VariableToken
-} from "sparqler/tokens";
-
-import { createMockDigestedSchemaProperty } from "../../test/helpers/mocks";
-
-import { PointerType } from "../ObjectSchema/PointerType";
-
 import { hasSignature, INSTANCE, method, module } from "../test/JasmineExtender";
-
-import { XSD } from "../Vocabularies/XSD";
-
-import { QueryContext } from "./QueryContext";
 
 import * as Utils from "./Utils";
 
@@ -28,708 +10,7 @@ describe( module( "carbonldp/QueryDocuments/Utils" ), ():void => {
 		expect( Utils ).toEqual( jasmine.any( Object ) );
 	} );
 
-	describe( method( INSTANCE, "_getLevelRegExp" ), ():void => {
-
-		it( hasSignature(
-			"Creates a regex to match child properties by name since an specific property",
-			[
-				{ name: "property", type: "string", description: "The name of the property to create the regex from." },
-			],
-			{ type: "RegExp" }
-		), ():void => {
-		} );
-
-		it( "should exists", ():void => {
-			expect( Utils._getLevelRegExp ).toBeDefined();
-			expect( Utils._getLevelRegExp ).toEqual( jasmine.any( Function ) );
-		} );
-
-		it( "should create a regex for no property level", ():void => {
-			const regex:RegExp = Utils._getLevelRegExp( "" );
-			expect( regex ).toEqual( /^[^.]+$/ );
-		} );
-
-		it( "should create a regex for a specific property level", ():void => {
-			const documentLevel:RegExp = Utils._getLevelRegExp( "document" );
-			expect( documentLevel ).toEqual( /^document\.[^.]+$/ );
-
-			const propertyLevel:RegExp = Utils._getLevelRegExp( "property" );
-			expect( propertyLevel ).toEqual( /^property\.[^.]+$/ );
-		} );
-
-		it( "should create a regex for a specific sub property level", ():void => {
-			const subPropertyLevel:RegExp = Utils._getLevelRegExp( "property.sub-property" );
-			expect( subPropertyLevel ).toEqual( /^property\.sub-property\.[^.]+$/ );
-
-			const subSubPropertyLevel:RegExp = Utils._getLevelRegExp( "property.sub-property.sub-sub-property" );
-			expect( subSubPropertyLevel ).toEqual( /^property\.sub-property\.sub-sub-property\.[^.]+$/ );
-		} );
-
-	} );
-
-
-	describe( method( INSTANCE, "_createPropertyPatterns" ), ():void => {
-
-		it( hasSignature(
-			"Creates the base patterns for a query property.",
-			[
-				{ name: "context", type: "CarbonLDP.QueryDocuments.QueryContext", description: "The context of the query where to obtains the information for the patterns." },
-				{ name: "resourcePath", type: "string", description: "Path of the parent property/resource where the property belongs to." },
-				{ name: "propertyPath", type: "string", description: "Path of the property to create its patterns." },
-				{ name: "propertyDefinition", type: "CarbonLDP.DigestedObjectSchemaProperty", description: "Schema definition of the property." },
-			],
-			{ type: "SPARQLER/tokens/PatternToken[]" }
-		), ():void => {
-		} );
-
-		it( "should exists", ():void => {
-			expect( Utils._createPropertyPatterns ).toBeDefined();
-			expect( Utils._createPropertyPatterns ).toEqual( jasmine.any( Function ) );
-		} );
-
-
-		let context:QueryContext;
-		beforeEach( ():void => {
-			context = new QueryContext();
-		} );
-
-		it( "should return a subject pattern", () => {
-			const patterns:PatternToken[] = Utils._createPropertyPatterns(
-				context,
-				"resource",
-				"property",
-				createMockDigestedSchemaProperty( { uri: "https://example.com/ns#property" } )
-			);
-
-			expect<any>( patterns ).toEqual( [
-				jasmine.objectContaining( {
-					token: "subject",
-					subject: jasmine.objectContaining( {
-						token: "variable",
-						name: "resource",
-					} ),
-					properties: [
-						jasmine.objectContaining( {
-							token: "property",
-							verb: jasmine.objectContaining( {
-								token: "iri",
-								value: "https://example.com/ns#property",
-							} ),
-							objects: [
-								jasmine.objectContaining( {
-									token: "variable",
-									name: "property",
-								} ),
-							],
-						} ),
-					],
-				} ),
-			] );
-		} );
-
-		it( "should return a filter datatype when definition specified a literal type", () => {
-			const patterns:PatternToken[] = Utils._createPropertyPatterns(
-				context,
-				"resource",
-				"property",
-				createMockDigestedSchemaProperty( {
-					uri: "https://example.com/ns#property",
-					literalType: XSD.boolean,
-				} )
-			);
-
-			expect<any>( patterns ).toEqual( [
-				jasmine.objectContaining( {
-					token: "subject",
-				} ),
-				jasmine.objectContaining( {
-					token: "filter",
-					constraint: `datatype( ?property ) = <${ XSD.boolean }>`,
-				} ),
-			] );
-		} );
-
-		it( "should return a filter not literals when definition specified a pointer type", () => {
-			const patterns:PatternToken[] = Utils._createPropertyPatterns(
-				context,
-				"resource",
-				"property",
-				createMockDigestedSchemaProperty( {
-					uri: "https://example.com/ns#property",
-					pointerType: PointerType.ID,
-				} )
-			);
-
-			expect<any>( patterns ).toEqual( [
-				jasmine.objectContaining( {
-					token: "subject",
-				} ),
-				jasmine.objectContaining( {
-					token: "filter",
-					constraint: `! isLiteral( ?property )`,
-				} ),
-			] );
-		} );
-
-	} );
-
-	describe( method( INSTANCE, "_createTypesPattern" ), ():void => {
-
-		it( hasSignature(
-			"Creates the pattern to query the types of a property/resource",
-			[
-				{ name: "context", type: "CarbonLDP.QueryDocuments.QueryContext", description: "The context of the query where to obtains the information for the patterns." },
-				{ name: "resourcePath", type: "string", description: "Path of the parent property/resource to query its types." },
-			],
-			{ type: "SPARQLER/tokens/PatternToken" }
-		), ():void => {
-		} );
-
-		it( "should exists", ():void => {
-			expect( Utils._createTypesPattern ).toBeDefined();
-			expect( Utils._createTypesPattern ).toEqual( jasmine.any( Function ) );
-		} );
-
-
-		let context:QueryContext;
-		beforeEach( ():void => {
-			context = new QueryContext();
-		} );
-
-		it( "should return a pattern token for types", () => {
-			const pattern:PatternToken = Utils._createTypesPattern( context, "resource" );
-			expect<any>( pattern ).toEqual( jasmine.objectContaining( {
-				token: "optional" as "optional",
-				groupPattern: jasmine.objectContaining( {
-					token: "groupPattern" as "groupPattern",
-					patterns: [
-						jasmine.objectContaining( {
-							token: "subject" as "subject",
-							subject: jasmine.objectContaining( {
-								token: "variable" as "variable",
-								name: "resource",
-							} ),
-							properties: [
-								jasmine.objectContaining( {
-									token: "property" as "property",
-									verb: "a",
-									objects: [
-										jasmine.objectContaining( {
-											token: "variable" as "variable",
-											name: "resource__types",
-										} ),
-									],
-								} ),
-							],
-						} ),
-					],
-				} ),
-			} ) );
-		} );
-
-	} );
-
-	describe( method( INSTANCE, "_createGraphPattern" ), ():void => {
-
-		it( hasSignature(
-			"Creates the graph pattern of a complete query.",
-			[
-				{ name: "context", type: "CarbonLDP.QueryDocuments.QueryContext", description: "The context of the query where to obtains the information for the patterns." },
-				{ name: "resourcePath", type: "string", description: "Path of the parent property/resource to create graph pattern." },
-			],
-			{ type: "SPARQLER/tokens/PatternToken" }
-		), ():void => {
-		} );
-
-		it( "should exists", ():void => {
-			expect( Utils._createTypesPattern ).toBeDefined();
-			expect( Utils._createTypesPattern ).toEqual( jasmine.any( Function ) );
-		} );
-
-
-		let context:QueryContext;
-		beforeEach( ():void => {
-			context = new QueryContext();
-		} );
-
-		it( "should return a pattern token for the resource", () => {
-			const pattern:PatternToken = Utils._createGraphPattern( context, "resource" );
-			expect<any>( pattern ).toEqual( jasmine.objectContaining( {
-				token: "graph" as "graph",
-				graph: jasmine.objectContaining( {
-					token: "variable" as "variable",
-					name: "resource",
-				} ),
-				groupPattern: jasmine.objectContaining( {
-					token: "groupPattern" as "groupPattern",
-					patterns: [
-						jasmine.objectContaining( {
-							token: "subject" as "subject",
-							subject: jasmine.objectContaining( {
-								token: "variable" as "variable",
-								name: "resource___subject",
-							} ),
-							properties: [
-								jasmine.objectContaining( {
-									token: "property" as "property",
-									verb: jasmine.objectContaining( {
-										token: "variable" as "variable",
-										name: "resource___predicate",
-									} ),
-									objects: [
-										jasmine.objectContaining( {
-											token: "variable" as "variable",
-											name: "resource___object",
-										} ),
-									],
-								} ),
-							],
-						} ),
-					],
-				} ),
-			} ) );
-		} );
-
-	} );
-
-	describe( method( INSTANCE, "_createAllPattern" ), ():void => {
-
-		it( hasSignature(
-			"Creates the pattern of return the all the properties of a resource (only the first level).",
-			[
-				{ name: "context", type: "CarbonLDP.QueryDocuments.QueryContext", description: "The context of the query where to obtains the information for the patterns." },
-				{ name: "resourcePath", type: "string", description: "Path of the parent property/resource to create the pattern." },
-			],
-			{ type: "SPARQLER/tokens/PatternToken" }
-		), ():void => {} );
-
-		it( "should exists", ():void => {
-			expect( Utils._createAllPattern ).toBeDefined();
-			expect( Utils._createAllPattern ).toEqual( jasmine.any( Function ) );
-		} );
-
-
-		let context:QueryContext;
-		beforeEach( ():void => {
-			context = new QueryContext();
-		} );
-
-		it( "should return a pattern token for the resource", () => {
-			const pattern:PatternToken = Utils._createAllPattern( context, "resource" );
-			expect<any>( pattern ).toEqual( jasmine.objectContaining( {
-				token: "subject" as "subject",
-				subject: jasmine.objectContaining( {
-					token: "variable" as "variable",
-					name: "resource",
-				} ),
-				properties: [
-					jasmine.objectContaining( {
-						token: "property" as "property",
-						verb: jasmine.objectContaining( {
-							token: "variable" as "variable",
-							name: "resource___predicate",
-						} ),
-						objects: [
-							jasmine.objectContaining( {
-								token: "variable" as "variable",
-								name: "resource___object",
-							} ),
-						],
-					} ),
-				],
-			} ), );
-		} );
-
-	} );
-
-
-	describe( method( INSTANCE, "_getParentPath" ), ():void => {
-
-		it( hasSignature(
-			"Returns the path for the immediately parent for the one provided.",
-			[
-				{ name: "path", type: "string", description: "Path to obtains its parent path." },
-			],
-			{ type: "string" }
-		), ():void => {} );
-
-		it( "should exists", ():void => {
-			expect( Utils._getParentPath ).toBeDefined();
-			expect( Utils._getParentPath ).toEqual( jasmine.any( Function ) );
-		} );
-
-
-		it( "should return empty when no parent", () => {
-			const path:string = Utils._getParentPath( "resource" );
-			expect( path ).toBe( "" );
-		} );
-
-		it( "should return the parent path when only one parent", () => {
-			const path:string = Utils._getParentPath( "parent.resource" );
-			expect( path ).toBe( "parent" );
-		} );
-
-		it( "should return the parent path when more that one parent", () => {
-			const path:string = Utils._getParentPath( "parent3.parent2.parent1.resource" );
-			expect( path ).toBe( "parent3.parent2.parent1" );
-		} );
-
-	} );
-
-
-	describe( method( INSTANCE, "_getAllTriples" ), () => {
-
-		it( hasSignature(
-			"Search and return all the triples variables (compacted in same subject) in all the levels of the patterns provided.",
-			[
-				{ name: "patterns", type: "SPARQLER/tokens/PatternToken", description: "Patterns where to look for the triples." },
-			],
-			{ type: "SPARQLER/tokens/SubjectToken" }
-		), ():void => {} );
-
-		it( "should exists", ():void => {
-			expect( Utils._getAllTriples ).toBeDefined();
-			expect( Utils._getAllTriples ).toEqual( jasmine.any( Function ) );
-		} );
-
-
-		it( "should not return subject with no variable", () => {
-			const patterns:PatternToken[] = [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new IRIRefToken( "https://example.com/ns#Type" ) )
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [] );
-		} );
-
-		it( "should return same subject when valid", () => {
-			const patterns:PatternToken[] = [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-			] );
-		} );
-
-		it( "should return multiple valid triples", () => {
-			const patterns:PatternToken[] = [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-				new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-				new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-			] );
-		} );
-
-
-		it( "should return subject from inside optional", () => {
-			const patterns:PatternToken[] = [
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-			] );
-		} );
-
-		it( "should multiple subjects from inside optionals", () => {
-			const patterns:PatternToken[] = [
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-						.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-							.addObject( new VariableToken( "property" ) )
-						)
-					)
-					.addPattern( new SubjectToken( new VariableToken( "property" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "property_types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-				new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-				new SubjectToken( new VariableToken( "property" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "property_types" ) )
-					),
-			] );
-		} );
-
-		it( "should return subject from inside graph", () => {
-			const patterns:PatternToken[] = [
-				new GraphToken( new IRIRefToken( "https://example.com/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-			] );
-		} );
-
-		it( "should multiple subjects from inside graphs", () => {
-			const patterns:PatternToken[] = [
-				new GraphToken( new IRIRefToken( "https://example.com/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-				new GraphToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-						.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-							.addObject( new VariableToken( "property" ) )
-						)
-					)
-					.addPattern( new SubjectToken( new VariableToken( "property" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "property_types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-				new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-				new SubjectToken( new VariableToken( "property" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "property_types" ) )
-					),
-			] );
-		} );
-
-		it( "should multiple subjects from inside a optional and graphs", () => {
-			const patterns:PatternToken[] = [
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-				new GraphToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-						.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-							.addObject( new VariableToken( "property" ) )
-						)
-					)
-					.addPattern( new SubjectToken( new VariableToken( "property" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "property_types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-				new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-				new SubjectToken( new VariableToken( "property" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "property_types" ) )
-					),
-			] );
-		} );
-
-		it( "should multiple subjects from inside a inside graphs inside optionals", () => {
-			const patterns:PatternToken[] = [
-				new GraphToken( new IRIRefToken( "https://example.com/" ) )
-					.addPattern( new OptionalToken()
-						.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-							.addProperty( new PropertyToken( "a" )
-								.addObject( new VariableToken( "types" ) )
-							)
-						)
-					),
-				new OptionalToken()
-					.addPattern( new GraphToken( new IRIRefToken( "https://example.com/resource/" ) )
-						.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-							.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-								.addObject( new VariableToken( "property" ) )
-							)
-						)
-						.addPattern( new SubjectToken( new VariableToken( "property" ) )
-							.addProperty( new PropertyToken( "a" )
-								.addObject( new VariableToken( "property_types" ) )
-							)
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					),
-				new SubjectToken( new IRIRefToken( "https://example.com/resource/" ) )
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-				new SubjectToken( new VariableToken( "property" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "property_types" ) )
-					),
-			] );
-		} );
-
-
-		it( "should combine into same subject when in different sub-patterns", () => {
-			const patterns:PatternToken[] = [
-				new GraphToken( new IRIRefToken( "https://example.com/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-							.addObject( new VariableToken( "property" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new VariableToken( "types" ) )
-					)
-					.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-						.addObject( new VariableToken( "property" ) )
-					),
-			] );
-		} );
-
-		it( "should only return subject with predicate as variable when first subject", () => {
-			const patterns:PatternToken[] = [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( new VariableToken( "predicates" ) )
-						.addObject( new VariableToken( "values" ) )
-					),
-				new GraphToken( new IRIRefToken( "https://example.com/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( new VariableToken( "predicates" ) )
-						.addObject( new VariableToken( "values" ) )
-					),
-			] );
-		} );
-
-		it( "should only return subject with predicate as variable when not first", () => {
-			const patterns:PatternToken[] = [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( "a" )
-						.addObject( new IRIRefToken( "https://example.com/ns#Type" ) )
-					),
-				new GraphToken( new IRIRefToken( "https://example.com/" ) )
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( new IRIRefToken( "https://example.com/ns#property" ) )
-							.addObject( new VariableToken( "property" ) )
-						)
-					),
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( new VariableToken( "predicates" ) )
-							.addObject( new VariableToken( "values" ) )
-						)
-					),
-				new OptionalToken()
-					.addPattern( new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-						.addProperty( new PropertyToken( "a" )
-							.addObject( new VariableToken( "types" ) )
-						)
-					),
-			];
-
-			const returned:PatternToken[] = Utils._getAllTriples( patterns );
-			expect( returned ).toEqual( [
-				new SubjectToken( new IRIRefToken( "https://example.com/" ) )
-					.addProperty( new PropertyToken( new VariableToken( "predicates" ) )
-						.addObject( new VariableToken( "values" ) )
-					),
-			] );
-		} );
-
-	} );
+	// TODO: Test ._getRootPath
 
 
 	describe( method( INSTANCE, "_getPathProperty" ), () => {
@@ -859,4 +140,9 @@ describe( module( "carbonldp/QueryDocuments/Utils" ), ():void => {
 		} );
 
 	} );
+
+
+	// TODO: Test ._getBestType
+	// TODO: Test ._getMatchingDefinition
+
 } );
